[BUGFIX][WORKAROUND] enable permission check for filemount table
the disabled check for filemounts makes it impossible to add / delete a filemount even as admin.
We do not know why this was titled as "bugged" from the extension author.
We tested the behavior with the enabled permission check for filemounts and found no problems.

Relates: https://projekte.in2code.de/issues/74288